README.md
The Sleep Drowsiness Detection System is a project designed to monitor and detect drowsiness in individuals, especially drivers, to prevent accidents caused by falling asleep at the wheel. The system uses Python along with computer vision and image processing techniques to track the user's eye movements and facial expressions in real-time.

This project is highly beneficial in the field of road safety, providing a practical application of Python and computer vision techniques to save lives.